chore: refactoring dto
