+++ src/Model/Dto/Image/StoredImageDto.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Model\Dto;
+namespace App\Model\Dto\Image;
 
 class StoredImageDto
 {
