+++ src/Model/Dto/Post/IndexPostDto.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Model\Dto;
+namespace App\Model\Dto\Post;
 
 class IndexPostDto
 {
