+++ src/Model/Dto/Post/ShowPostDto.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Model\Dto;
+namespace App\Model\Dto\Post;
 
 class ShowPostDto
 {
