+++ src/Model/Dto/Post/UpdatePostDto.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Model\Dto;
+namespace App\Model\Dto\Post;
 
 class UpdatePostDto
 {
