+++ src/Model/Dto/Tag/IndexTagDto.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Model\Dto;
+namespace App\Model\Dto\Tag;
 
 class IndexTagDto
 {
