--- src/Repository/PostCategoryRepositoryInterface.php
@@ -1,6 +1,0 @@
-<?php
-namespace App\Repository;
-
-interface PostCategoryRepositoryInterface {
-    public function insertMultiCategory(int $post_id, array $category_list);
-}
