+++ src/Repository/PostRepositoryInterface.php
@@ -2,10 +2,10 @@
 
 namespace App\Repository;
 
-use App\Model\Dto\DetailPostDto;
-use App\Model\Dto\IndexPostDto;
-use App\Model\Dto\ShowPostDto;
-use App\Model\Dto\UpdatePostDto;
+use App\Model\Dto\Post\DetailPostDto;
+use App\Model\Dto\Post\IndexPostDto;
+use App\Model\Dto\Post\ShowPostDto;
+use App\Model\Dto\Post\UpdatePostDto;
 
 interface PostRepositoryInterface
 {
