+++ src/Repository/TagRepositoryInterface.php
@@ -1,7 +1,7 @@
 <?php
 namespace App\Repository;
 
-use App\Model\Dto\IndexTagDto;
+use App\Model\Dto\Tag\IndexTagDto;
 
 interface TagRepositoryInterface {
 
@@ -15,4 +15,6 @@ interface TagRepositoryInterface {
      * @return IndexTagDto[]
      */
     public function getTagList(): array;
+
+    public function insertMultiTag(int $post_id, array $tag_list): void;
 }
