+++ src/Route.php
@@ -14,7 +14,6 @@ use App\Lib\Http\Request;
 use App\Lib\Singleton\PageCompose;
 use App\Lib\Singleton\PgConnect;
 use App\Repository\ImageRepository;
-use App\Repository\PostCategoryRepository;
 use App\Repository\PostRepository;
 use App\Repository\TagRepository;
 
@@ -29,7 +28,7 @@ class Route
         } else if ($req->method === "POST" && $req->path === "/") {
             session_start();
             $pdo = PgConnect::getClient();
-            return new CreatePosthandler(req: $req, pdo: $pdo, compose: PageCompose::getComposer(), post_repo: new PostRepository(pdo: $pdo), image_repo: new ImageRepository(pdo: $pdo), post_category_repo: new PostCategoryRepository(pdo: $pdo));
+            return new CreatePosthandler(req: $req, pdo: $pdo, compose: PageCompose::getComposer(), post_repo: new PostRepository(pdo: $pdo), image_repo: new ImageRepository(pdo: $pdo), tag_repo: new TagRepository(pdo: $pdo));
 
         } else if ($req->method === "GET" && preg_match("|\A/posts/([0-9]+)\z|u", $req->path, $match)) {
             $post_id = (int)$match[1];
